test(options): migrate rest compatibility tests
